+++ src/features/toolbar/ui/toolbar-no-auth.tsx
@@ -1,0 +1,17 @@
+import { Link } from 'react-router-dom';
+import { AppRoute } from '../../../shared/enum';
+
+const ToolbarNoAuth = (): JSX.Element => (
+  <li className="header__nav-item user">
+    <Link
+      className="header__nav-link header__nav-link--profile"
+      to={AppRoute.Login}
+    >
+      <div className="header__avatar-wrapper user__avatar-wrapper">
+      </div>
+      <span className="header__login">Sign in</span>
+    </Link>
+  </li>
+);
+
+export default ToolbarNoAuth;
